refac: create alloc file
