+++ includes/alloc.h
@@ -1,7 +1,9 @@
 #ifndef ALLOC_H
-#define ALLOC_H
+# define ALLOC_H
 
-#include <stddef.h>
+# include <stddef.h>
+
+void	*allocate_chunk(size_t size);
 
 void	*map_pages_by_size(size_t map_size);
