+++ srcs/alloc.c
@@ -1,0 +1,79 @@
+#include "ft_printf.h"
+#include "malloc.h"
+#include "utils.h"
+#include "alloc.h"
+#include <unistd.h>
+#include <stdio.h>
+
+t_malloc	g_regions;
+
+void	split_chunk(t_heap_chunk *chunk, size_t chunk_size)
+{
+	t_heap_chunk	*next;
+
+	next = (void *)chunk + chunk_size;
+	next->size = chunk->size - chunk_size;
+	chunk->size = chunk_size | IS_PREV_IN_USE(chunk);
+}
+
+void	*find_chunk_from_region(t_region *region, size_t chunk_size)
+{
+	t_heap_chunk	*chunk;
+
+	chunk = region->freelist;
+	while (chunk->fd && SIZE(chunk) < chunk_size)
+		chunk = chunk->fd;
+	if (SIZE(chunk) < chunk_size)
+	{
+		ft_printf("error not enough space\n");
+		return (NULL);
+	}
+	return (chunk);
+}
+
+void	*allocate_chunk_from_region(t_region *region, size_t size)
+{
+	size_t			chunk_size;
+	t_heap_chunk	*chunk;
+
+	chunk_size = align_chunk_size(size);
+	chunk = find_chunk_from_region(region, chunk_size);
+	if (!chunk)
+		return (NULL);
+	if (SIZE(chunk) > chunk_size)
+	{
+		split_chunk(chunk, chunk_size);
+		freelst_replace(chunk, NEXTCHUNK(chunk), &(region->freelist));
+	}
+	else
+		freelst_pop(chunk, &(region->freelist));
+	chunk->size |= ALLOCED;
+	NEXTCHUNK(chunk)->size |= PREV_IN_USE;
+	return (MEM(chunk));
+}
+
+void	*allocate_chunk_from_heap(size_t size)
+{
+	t_heap_chunk	*chunk;
+	int				page_size;
+
+	page_size = get_page_size();
+	if (page_size == ERROR)
+		return (NULL);
+	size = align(size, page_size);
+	chunk = map_pages_by_size(size);
+	chunk->size = size | MAPPED;
+	return (MEM(chunk));
+}
+
+void	*allocate_chunk(size_t size)
+{
+	size_t	aligned_size;
+
+	aligned_size = align_size(size);
+	if (aligned_size < TINY_MAX)
+		return (allocate_chunk_from_region(&(g_regions.tiny_region), aligned_size));
+	else if (aligned_size < SMALL_MAX)
+		return (allocate_chunk_from_region(&(g_regions.small_region), aligned_size));
+	return (allocate_chunk_from_heap(size));
+}
